♻️ Separate verify steps to utils functions for CCU
